-0x8000 == 0x8000 with int16 fixes segfault / issue491
Originally committed as revision 13767 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavcodec/g726.c
@@ -37,7 +37,7 @@ typedef struct Float11 {
     int mant;   /**< 6bit mantissa */
 } Float11;
 
-static inline Float11* i2f(int16_t i, Float11* f)
+static inline Float11* i2f(int i, Float11* f)
 {
     f->sign = (i < 0);
     if (f->sign)
